requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
django-stub==0.1
django-stubs==1.5.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
djoser==2.0.3
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mypy==0.770
mypy-extensions==0.4.3
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
